requirements.txt
librosa
matplotlib
numpy
pillow
soundfile
